% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_screening.R
\name{cor_screening}
\alias{cor_screening}
\title{cor_screening}
\usage{
cor_screening(X, nlambda = NULL, lambda_min_ratio = NULL)
}
\arguments{
\item{X}{a n by p data matrix containing iid observations. Each row is a new sample.}

\item{nlambda}{the number of tuning parameters. Default \code{NULL}.}

\item{lambda_min_ratio}{the largest sparsity level of the sparse adjacency matrix estimates. Default value \code{0.05}.}
}
\value{
A list containing array of sparse adjacency matrix estimates, and corresponding tuning parameter values.
}
\description{
Computes a path of sparse adjacency matrix estimates using correlation thresholding
}
\examples{

library(huge)
library(igraph)
library(MASS)

n <- 100
p <- 50

Data <- huge::huge.generator(n = n, d = p, graph = "hub")

true_nmb_hubs <- ifelse(p > 40, ceiling(p/20), 2)

Sigma <- Data$sigma

A_true <- Data$theta

A_true <- as.matrix(A_true)

true_hubs <- colSums(A_true)

true_hubs <- order(-true_hubs)[1:true_nmb_hubs]

A_true[A_true != 0] <- 1

diag(A_true) <- 0

G_true <- igraph::graph_from_adjacency_matrix(A_true, mode = "undirected", diag = FALSE)

node_names <- as.character(1:p)

vertex_attr(G_true) <- list(name = node_names)

d_true <- igraph::degree(G_true)

degree_true <- colSums(A_true)

names(degree_true) <- 1:p

X = MASS::mvrnorm(n = n, mu = rep(0, p), Sigma)

X = scale(X)

nlambda = 70

cor_path = cor_screening(X = X, nlambda = nlambda)

}
