% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_detection_mdsd.R
\name{hub_detection_mdsd}
\alias{hub_detection_mdsd}
\title{hub_detection_mdsd}
\usage{
hub_detection_mdsd(
  sol_path = NULL,
  node_names = NULL,
  gamma = NULL,
  skew_thr = NULL
)
}
\arguments{
\item{sol_path}{a list containing an array \code{wi} of spare matrices and the corresponding tuning parameter values \code{rho_list}.}

\item{node_names}{node names. If not provided, nodes are numbered from 1 to \code{ncol(sol_path$wi)}. Default \code{NULL}.}

\item{gamma}{the multiplier of the 'outlier' cut-off value. Default value \code{3}.}

\item{skew_thr}{the skewness threshold. Only a subset of models with estimated degree distribution skewness greater than \code{skew_thr} are inspected. Default value \code{0.5}.}
}
\value{
A list of potential hub nodes, MDSD distance values, and node degrees.
}
\description{
Computes the MDSD distance from the solution path information.
}
\examples{

library(huge)
library(igraph)
library(MASS)

n <- 100
p <- 50

Data <- huge::huge.generator(n = n, d = p, graph = "hub")

true_nmb_hubs <- ifelse(p > 40, ceiling(p/20), 2)

Sigma <- Data$sigma

A_true <- Data$theta

A_true <- as.matrix(A_true)

true_hubs <- colSums(A_true)

true_hubs <- order(-true_hubs)[1:true_nmb_hubs]

A_true[A_true != 0] <- 1

diag(A_true) <- 0

G_true <- igraph::graph_from_adjacency_matrix(A_true, mode = "undirected", diag = FALSE)

node_names <- as.character(1:p)

vertex_attr(G_true) <- list(name = node_names)

d_true <- igraph::degree(G_true)

degree_true <- colSums(A_true)

names(degree_true) <- 1:p

X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma)

X <- scale(X)

nlambda <- 70

cor_path <- cor_screening(X = X, nlambda = nlambda)

res <- hub_detection_mdsd(sol_path = cor_path)

}
